feat(FR-1749): Admin-dedicated sidebar menu (#4791)
Resolves #4739 ([FR-1749](https://lablup.atlassian.net/browse/FR-1749))

This is a re-written PR for the purpose of replacing the PR https://github.com/lablup/backend.ai-webui/pull/4741. For related matters, please refer to https://github.com/lablup/backend.ai-webui/pull/4741 and FR-1749.

1. Refactoring the WebUI plugin state management into a dedicated hook (`useWebUIPluginState`)
2. Moving menu item definitions to a new hook (`useWebUIMenuItems`) for better organization
3. Adding a "Go Back" button in the admin settings menu to return to the last visited general page
4. Implementing a cleaner UI for admin settings with a dedicated header
5. Storing the last visited general menu path in session storage for navigation history
6. Adding translations for new UI elements across all supported languages

The changes improve the user experience by making it easier to navigate between admin settings and general pages.

[FR-1749]: https://lablup.atlassian.net/browse/FR-1749?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
