+++ .cspell.json
@@ -26,6 +26,7 @@
     "Backendai",
     "backendaiclient",
     "backendaioptions",
+    "backendaiwebui",
     "baseversion",
     "cooldown",
     "cssinjs",
