+++ resources/i18n/de.json
@@ -2084,6 +2084,7 @@
       "AIAgents": "AI-Agenten",
       "AboutBackendAI": "Über Backend.AI",
       "AdminDashboard": "Admin -Dashboard",
+      "AdminSettings": "Admin-Einstellungen",
       "Administration": "Verwaltung",
       "AgentSummary": "Agent Zusammenfassung",
       "Architecture": "Architektur",
@@ -2110,6 +2111,7 @@
       "FullName": "Vollständiger Name",
       "FullNameInvalid": "Der vollständige Name sollte 64 Zeichen nicht überschreiten",
       "FullNameUpdated": "Der vollständige Name wurde erfolgreich aktualisiert.",
+      "GoBack": "Zurück",
       "Help": "Hilfe",
       "Import&Run": "Importieren und ausführen",
       "Information": "Information",
