+++ resources/i18n/el.json
@@ -2082,6 +2082,7 @@
       "AIAgents": "Πράκτορες AI",
       "AboutBackendAI": "Σχετικά με το Backend.AI",
       "AdminDashboard": "Πίνακας ελέγχου διαχειριστή",
+      "AdminSettings": "Ρυθμίσεις διαχειριστή",
       "Administration": "Διαχείριση",
       "AgentSummary": "Περίληψη πράκτορα",
       "Architecture": "Αρχιτεκτονική",
@@ -2107,6 +2108,7 @@
       "FullName": "Πλήρες όνομα",
       "FullNameInvalid": "Το πλήρες όνομα δεν πρέπει να υπερβαίνει τους 64 χαρακτήρες",
       "FullNameUpdated": "Το πλήρες όνομα ενημερώθηκε με επιτυχία.",
+      "GoBack": "Πίσω",
       "Help": "Βοήθεια",
       "Import&Run": "Εισαγωγή και εκτέλεση",
       "Information": "Πληροφορίες",
