+++ resources/i18n/en.json
@@ -2093,6 +2093,7 @@
       "AIAgents": "AI Agents",
       "AboutBackendAI": "About Backend.AI",
       "AdminDashboard": "Admin Dashboard",
+      "AdminSettings": "Admin Settings",
       "Administration": "Administration",
       "AgentSummary": "Agent Summary",
       "Architecture": "Architecture",
@@ -2119,6 +2120,7 @@
       "FullName": "Full Name",
       "FullNameInvalid": "Full name should not exceed 64 characters",
       "FullNameUpdated": "Full name is successfully updated.",
+      "GoBack": "Go Back",
       "Help": "Help",
       "Import&Run": "Import & Run",
       "Information": "Information",
