+++ resources/i18n/es.json
@@ -2086,6 +2086,7 @@
       "AIAgents": "Agentes de IA",
       "AboutBackendAI": "Acerca de Backend.AI",
       "AdminDashboard": "Tablero de administración",
+      "AdminSettings": "Ajustes de administrador",
       "Administration": "Administración",
       "AgentSummary": "Resumen del agente",
       "Architecture": "Arquitectura",
@@ -2112,6 +2113,7 @@
       "FullName": "Nombre y apellidos",
       "FullNameInvalid": "El nombre completo no debe superar los 64 caracteres",
       "FullNameUpdated": "El nombre completo se ha actualizado correctamente.",
+      "GoBack": "Atrás",
       "Help": "Ayuda",
       "Import&Run": "Importar y ejecutar",
       "Information": "Información",
