+++ resources/i18n/fi.json
@@ -2085,6 +2085,7 @@
       "AIAgents": "AI-agentit",
       "AboutBackendAI": "Tietoja Backend.AI:sta",
       "AdminDashboard": "Järjestelmänvalvojan kojelauta",
+      "AdminSettings": "Ylläpitäjän asetukset",
       "Administration": "Hallinto",
       "AgentSummary": "Agentin yhteenveto",
       "Architecture": "Arkkitehtuuri",
@@ -2111,6 +2112,7 @@
       "FullName": "Koko nimi",
       "FullNameInvalid": "Koko nimi saa olla enintään 64 merkkiä",
       "FullNameUpdated": "Koko nimi on päivitetty onnistuneesti.",
+      "GoBack": "Takaisin",
       "Help": "Apua",
       "Import&Run": "Tuo ja suorita",
       "Information": "Tiedot",
