+++ resources/i18n/fr.json
@@ -2086,6 +2086,7 @@
       "AIAgents": "Agents d'intelligence artificielle",
       "AboutBackendAI": "À propos de Backend.AI",
       "AdminDashboard": "Tableau de bord administratif",
+      "AdminSettings": "Paramètres d'administration",
       "Administration": "Administration",
       "AgentSummary": "Résumé de l'agent",
       "Architecture": "Architecture",
@@ -2112,6 +2113,7 @@
       "FullName": "Nom et prénom",
       "FullNameInvalid": "Le nom complet ne doit pas dépasser 64 caractères",
       "FullNameUpdated": "Le nom complet a été mis à jour avec succès.",
+      "GoBack": "Retour",
       "Help": "Aide",
       "Import&Run": "Importer et exécuter",
       "Information": "Informations",
