+++ resources/i18n/id.json
@@ -2085,6 +2085,7 @@
       "AIAgents": "Agen AI",
       "AboutBackendAI": "Tentang Backend.AI",
       "AdminDashboard": "Dasbor Admin",
+      "AdminSettings": "Pengaturan Admin",
       "Administration": "Administrasi",
       "AgentSummary": "Ringkasan Agen",
       "Architecture": "Arsitektur",
@@ -2111,6 +2112,7 @@
       "FullName": "Nama lengkap",
       "FullNameInvalid": "Nama lengkap tidak boleh melebihi 64 karakter",
       "FullNameUpdated": "Nama lengkap berhasil diperbarui.",
+      "GoBack": "Kembali",
       "Help": "Bantuan",
       "Import&Run": "Impor & Jalankan",
       "Information": "Informasi",
