+++ resources/i18n/it.json
@@ -2083,6 +2083,7 @@
       "AIAgents": "Agenti AI",
       "AboutBackendAI": "Informazioni su Backend.AI",
       "AdminDashboard": "Dashboard di amministrazione",
+      "AdminSettings": "Impostazioni amministratore",
       "Administration": "Amministrazione",
       "AgentSummary": "Riepilogo dell'agente",
       "Architecture": "Architettura",
@@ -2109,6 +2110,7 @@
       "FullName": "Nome e cognome",
       "FullNameInvalid": "Il nome completo non deve superare i 64 caratteri",
       "FullNameUpdated": "Il nome completo è stato aggiornato con successo.",
+      "GoBack": "Indietro",
       "Help": "Aiuto",
       "Import&Run": "Importa ed esegui",
       "Information": "Informazione",
