+++ resources/i18n/ja.json
@@ -2085,6 +2085,7 @@
       "AIAgents": "AIエージェント",
       "AboutBackendAI": "Backend.AIについて",
       "AdminDashboard": "管理ダッシュボード",
+      "AdminSettings": "管理者設定",
       "Administration": "管理",
       "AgentSummary": "リソース概要",
       "Architecture": "建築",
@@ -2111,6 +2112,7 @@
       "FullName": "フルネーム",
       "FullNameInvalid": "フルネームは64文字を超えてはなりません",
       "FullNameUpdated": "フルネームが正常に更新されました。",
+      "GoBack": "戻る",
       "Help": "案内(ヘルプ)",
       "Import&Run": "インポート & 実行",
       "Information": "情報",
