+++ resources/i18n/ko.json
@@ -2090,6 +2090,7 @@
       "AIAgents": "AI 에이전트",
       "AboutBackendAI": "Backend.AI에 대하여",
       "AdminDashboard": "관리자 대시보드",
+      "AdminSettings": "관리자 설정",
       "Administration": "관리",
       "AgentSummary": "자원 요약",
       "Architecture": "운영체제",
@@ -2116,6 +2117,7 @@
       "FullName": "사용자 이름",
       "FullNameInvalid": "사용자 이름은 64자 이하로 입력하십시오",
       "FullNameUpdated": "사용자 이름이 갱신되었습니다.",
+      "GoBack": "뒤로가기",
       "Help": "도움말",
       "Import&Run": "가져오기 & 실행",
       "Information": "정보",
