+++ resources/i18n/mn.json
@@ -2083,6 +2083,7 @@
       "AIAgents": "AI агентууд",
       "AboutBackendAI": "Backend.AI-ийн тухай",
       "AdminDashboard": "Админ хяналтын самбар",
+      "AdminSettings": "Администраторын тохиргоо",
       "Administration": "Админ",
       "AgentSummary": "Агентын хураангуй",
       "Architecture": "Уран барилгач",
@@ -2109,6 +2110,7 @@
       "FullName": "Бүтэн нэр",
       "FullNameInvalid": "Бүтэн нэр 64 тэмдэгтээс хэтрэхгүй байх ёстой",
       "FullNameUpdated": "Бүтэн нэр амжилттай шинэчлэгдлээ.",
+      "GoBack": "Буцах",
       "Help": "Туслаач",
       "Import&Run": "Импортлох & ажиллуулах",
       "Information": "Мэдээлэл",
