+++ resources/i18n/ms.json
@@ -2084,6 +2084,7 @@
       "AIAgents": "Ejen AI",
       "AboutBackendAI": "Mengenai Backend.AI",
       "AdminDashboard": "Papan pemuka admin",
+      "AdminSettings": "Tetapan Pentadbir",
       "Administration": "Pentadbiran",
       "AgentSummary": "Ringkasan Agen",
       "Architecture": "Seni bina",
@@ -2110,6 +2111,7 @@
       "FullName": "Nama penuh",
       "FullNameInvalid": "Nama penuh tidak boleh melebihi 64 aksara",
       "FullNameUpdated": "Nama penuh berjaya dikemas kini.",
+      "GoBack": "Kembali",
       "Help": "Tolonglah",
       "Import&Run": "Import & Jalankan",
       "Information": "Maklumat",
