+++ resources/i18n/pl.json
@@ -2085,6 +2085,7 @@
       "AIAgents": "Agenci AI",
       "AboutBackendAI": "O Backend.AI",
       "AdminDashboard": "Pulpit nawigacyjny administratora",
+      "AdminSettings": "Ustawienia admina",
       "Administration": "Administracja",
       "AgentSummary": "Podsumowanie agenta",
       "Architecture": "Architektura",
@@ -2111,6 +2112,7 @@
       "FullName": "Pełne imię i nazwisko",
       "FullNameInvalid": "Pełna nazwa nie powinna przekraczać 64 znaków",
       "FullNameUpdated": "Pełna nazwa została pomyślnie zaktualizowana.",
+      "GoBack": "Wstecz",
       "Help": "Wsparcie",
       "Import&Run": "Importuj i uruchom",
       "Information": "Informacja",
