+++ resources/i18n/pt.json
@@ -2085,6 +2085,7 @@
       "AIAgents": "Agentes de IA",
       "AboutBackendAI": "Sobre Backend.AI",
       "AdminDashboard": "Painel de administrador",
+      "AdminSettings": "Config. Admin",
       "Administration": "Administração",
       "AgentSummary": "Resumo do agente",
       "Architecture": "Arquitetura",
@@ -2111,6 +2112,7 @@
       "FullName": "Nome completo",
       "FullNameInvalid": "O nome completo não deve ter mais de 64 caracteres",
       "FullNameUpdated": "O nome completo foi atualizado com sucesso.",
+      "GoBack": "Voltar",
       "Help": "Ajuda",
       "Import&Run": "Importar e executar",
       "Information": "Em formação",
