+++ resources/i18n/ru.json
@@ -2085,6 +2085,7 @@
       "AIAgents": "Агенты искусственного интеллекта",
       "AboutBackendAI": "О Backend.AI",
       "AdminDashboard": "Административная панель",
+      "AdminSettings": "Настройки администратора",
       "Administration": "Администрация",
       "AgentSummary": "Резюме агента",
       "Architecture": "Архитектура",
@@ -2111,6 +2112,7 @@
       "FullName": "ФИО",
       "FullNameInvalid": "Полное имя не должно превышать 64 символа.",
       "FullNameUpdated": "Полное имя успешно обновлено.",
+      "GoBack": "Назад",
       "Help": "Справка",
       "Import&Run": "Импорт и запуск",
       "Information": "Информация",
