+++ resources/i18n/th.json
@@ -2070,6 +2070,7 @@
       "AIAgents": "ตัวแทน AI",
       "AboutBackendAI": "เกี่ยวกับ Backend.AI",
       "AdminDashboard": "แดชบอร์ดผู้ดูแลระบบ",
+      "AdminSettings": "การตั้งค่าผู้ดูแลระบบ",
       "Administration": "การจัดการระบบ",
       "AgentSummary": "สรุปตัวแทน",
       "Architecture": "สถาปัตยกรรม",
@@ -2096,6 +2097,7 @@
       "FullName": "ชื่อเต็ม",
       "FullNameInvalid": "ชื่อเต็มไม่ควรเกิน 64 ตัวอักษร",
       "FullNameUpdated": "อัปเดตชื่อเต็มเรียบร้อยแล้ว",
+      "GoBack": "ย้อนกลับ",
       "Help": "ช่วยเหลือ",
       "Import&Run": "นำเข้าและรัน",
       "Information": "ข้อมูล",
