+++ resources/i18n/tr.json
@@ -2086,6 +2086,7 @@
       "AIAgents": "Yapay Zeka Ajanları",
       "AboutBackendAI": "Backend.AI hakkında",
       "AdminDashboard": "Yönetici Gösterge Tablosu",
+      "AdminSettings": "Yönetici Ayarları",
       "Administration": "yönetim",
       "AgentSummary": "Temsilci Özeti",
       "Architecture": "Mimarlık",
@@ -2112,6 +2113,7 @@
       "FullName": "Ad Soyad",
       "FullNameInvalid": "Tam ad 64 karakteri geçmemelidir",
       "FullNameUpdated": "Tam ad başarıyla güncellendi.",
+      "GoBack": "Geri dön",
       "Help": "Yardım",
       "Import&Run": "İçe Aktar ve Çalıştır",
       "Information": "Bilgi",
