+++ resources/i18n/vi.json
@@ -2086,6 +2086,7 @@
       "AIAgents": "Đại lý AI",
       "AboutBackendAI": "Giới thiệu về Backend.AI",
       "AdminDashboard": "Bảng điều khiển quản trị viên",
+      "AdminSettings": "Cài đặt quản trị",
       "Administration": "Hành chính",
       "AgentSummary": "Tóm tắt đại lý",
       "Architecture": "Ngành kiến ​​​​trúc",
@@ -2112,6 +2113,7 @@
       "FullName": "Họ và tên",
       "FullNameInvalid": "Tên đầy đủ không được vượt quá 64 ký tự",
       "FullNameUpdated": "Tên đầy đủ được cập nhật thành công.",
+      "GoBack": "Quay lại",
       "Help": "Cứu giúp",
       "Import&Run": "Nhập & Chạy",
       "Information": "Thông tin",
