+++ resources/i18n/zh-CN.json
@@ -2086,6 +2086,7 @@
       "AIAgents": "人工智能代理",
       "AboutBackendAI": "关于后端.AI",
       "AdminDashboard": "管理仪表板",
+      "AdminSettings": "管理员设置",
       "Administration": "行政",
       "AgentSummary": "代理摘要",
       "Architecture": "建筑学",
@@ -2112,6 +2113,7 @@
       "FullName": "全名",
       "FullNameInvalid": "全名不应超过 64 个字符",
       "FullNameUpdated": "全名更新成功。",
+      "GoBack": "返回",
       "Help": "帮助",
       "Import&Run": "导入并运行",
       "Information": "信息",
