+++ resources/i18n/zh-TW.json
@@ -2084,6 +2084,7 @@
       "AIAgents": "人工智能代理",
       "AboutBackendAI": "關於後端.AI",
       "AdminDashboard": "管理儀表板",
+      "AdminSettings": "管理員設定",
       "Administration": "行政",
       "AgentSummary": "代理摘要",
       "Architecture": "建築學",
@@ -2110,6 +2111,7 @@
       "FullName": "全名",
       "FullNameInvalid": "全名不應超過 64 個字符",
       "FullNameUpdated": "全名更新成功。",
+      "GoBack": "返回",
       "Help": "幫助",
       "Import&Run": "導入並運行",
       "Information": "信息",
